change the way quality score is calculated, create migration to update existing overall score values

+++ API/RequirementAI.Contract/Dto/PersonaQualityScoreDto.cs
@@ -4,10 +4,10 @@ public class PersonaQualityScoreDto: QualityScoreBaseDto
 {
     public Guid PersonaId { get; set; }
 
-    public int ClarityScore { get; set; }
-    public int RealismScore { get; set; }
-    public int GoalClarityScore { get; set; }
-    public int PainPointsScore { get; set; }
-    public int RelevanceScore { get; set; }
-    public int DifferentiationScore { get; set; }
+    public decimal ClarityScore { get; set; }
+    public decimal RealismScore { get; set; }
+    public decimal GoalClarityScore { get; set; }
+    public decimal PainPointsScore { get; set; }
+    public decimal RelevanceScore { get; set; }
+    public decimal DifferentiationScore { get; set; }
 }
