+++ API/RequirementAI.Contract/Dto/UserStoryQualityScoreDto.cs
@@ -4,11 +4,11 @@ public class UserStoryQualityScoreDto: QualityScoreBaseDto
 {
     public Guid UserStoryId { get; set; }
 
-    public int ClarityScore { get; set; }
-    public int CompletenessScore { get; set; }
-    public int TestabilityScore { get; set; }
-    public int AcceptanceCriteriaScore { get; set; }
-    public int ScopeScore { get; set; }
-    public int BusinessValueScore { get; set; }
-    public int AmbiguityScore { get; set; }
+    public decimal ClarityScore { get; set; }
+    public decimal CompletenessScore { get; set; }
+    public decimal TestabilityScore { get; set; }
+    public decimal AcceptanceCriteriaScore { get; set; }
+    public decimal ScopeScore { get; set; }
+    public decimal BusinessValueScore { get; set; }
+    public decimal AmbiguityScore { get; set; }
 }
