+++ API/RequirementAI.Persistence/Entities/PersonaQualityScore.cs
@@ -5,10 +5,10 @@ public class PersonaQualityScore: QualityScoreBase
     public Guid PersonaId { get; set; }
     public Persona Persona { get; set; } = null!;
 
-    public int ClarityScore { get; set; }
-    public int RealismScore { get; set; }
-    public int GoalClarityScore { get; set; }
-    public int PainPointsScore { get; set; }
-    public int RelevanceScore { get; set; }
-    public int DifferentiationScore { get; set; }
+    public decimal ClarityScore { get; set; }
+    public decimal RealismScore { get; set; }
+    public decimal GoalClarityScore { get; set; }
+    public decimal PainPointsScore { get; set; }
+    public decimal RelevanceScore { get; set; }
+    public decimal DifferentiationScore { get; set; }
 }
