+++ API/RequirementAI.Persistence/Entities/ScenarioQualityScore.cs
@@ -5,10 +5,10 @@ public class ScenarioQualityScore: QualityScoreBase
     public Guid ScenarioId { get; set; }
     public Scenario Scenario { get; set; } = null!;
 
-    public int ClarityScore { get; set; }
-    public int ContextScore { get; set; }
-    public int TriggerScore { get; set; }
-    public int FlowCompletenessScore { get; set; }
-    public int EdgeCasesScore { get; set; }
-    public int PersonaFitScore { get; set; }
+    public decimal ClarityScore { get; set; }
+    public decimal ContextScore { get; set; }
+    public decimal TriggerScore { get; set; }
+    public decimal FlowCompletenessScore { get; set; }
+    public decimal EdgeCasesScore { get; set; }
+    public decimal PersonaFitScore { get; set; }
 }
